add if -e -js.pot

+++ gettext_po_generator_github.sh
@@ -63,16 +63,17 @@ if [ -n "$HTML_JS_FILES" ]; then
     rm $DIR/locale/$DIRNAME-tmp.pot
 
     # Combine files from bash and js/html
-    if [[ -e "$DIR/locale/$DIRNAME.pot" ]]; then
-        mv "$DIR/locale/$DIRNAME.pot" "$DIR/locale/$DIRNAME-bash.pot"
-        msgcat --no-wrap --strict "$DIR/locale/$DIRNAME-bash.pot" -i "$DIR/locale/$DIRNAME-js.pot" > $DIR/locale/$DIRNAME-tmp.pot
-        xgettext --package-name="$DIRNAME" --no-location -L PO -o "$DIR/locale/$DIRNAME.pot" -i "$DIR/locale/$DIRNAME-tmp.pot"
-        rm "$DIR/locale/$DIRNAME-bash.pot"
-        rm "$DIR/locale/$DIRNAME-js.pot"
-    else
-        mv "$DIR/locale/$DIRNAME-js.pot" "$DIR/locale/$DIRNAME.pot"
+    if [[ -e "$DIR/locale/$DIRNAME-js.pot" ]]; then
+        if [[ -e "$DIR/locale/$DIRNAME.pot" ]]; then
+            mv "$DIR/locale/$DIRNAME.pot" "$DIR/locale/$DIRNAME-bash.pot"
+            msgcat --no-wrap --strict "$DIR/locale/$DIRNAME-bash.pot" -i "$DIR/locale/$DIRNAME-js.pot" > $DIR/locale/$DIRNAME-tmp.pot
+            xgettext --package-name="$DIRNAME" --no-location -L PO -o "$DIR/locale/$DIRNAME.pot" -i "$DIR/locale/$DIRNAME-tmp.pot"
+            rm "$DIR/locale/$DIRNAME-bash.pot"
+            rm "$DIR/locale/$DIRNAME-js.pot"
+        else
+            mv "$DIR/locale/$DIRNAME-js.pot" "$DIR/locale/$DIRNAME.pot"
+        fi
     fi
-
 fi
 
 # Translate .py files
